requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typer==0.15.2
typing-extensions==4.13.1
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
